cast

+++ indexed_gzip/zran_file_util.c
@@ -87,7 +87,7 @@ fail:
  * This method requires the file size to be input, because there is no other way to tell if a Python file-like is at EOF without reading it.
  */
 int _feof_python(PyObject *f, uint64_t size) {
-    return _ftell_python(f) == size;
+    return _ftell_python(f) == (int64_t) size;
 }
 
 /*
